specific-scrapers/bheuu-scraper/pyproject.toml
[build-system]
requires = ["setuptools>=70"]
build-backend = "setuptools.backends.legacy:build"

[project]
name = "bheuu-scraper"
version = "1.0.0"
description = "Production-grade scraper for bheuu.gov.my – archives public documents to DigitalOcean Spaces for RAG pipelines"
requires-python = ">=3.9"
dependencies = [
    "requests>=2.31,<3",
    "python-dateutil>=2.9,<3",
    "boto3>=1.34,<2",
    "tenacity>=8.2,<9",
    "pyyaml>=6.0,<7",
    "click>=8.1,<9",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0,<9",
    "pytest-cov>=4.0,<6",
]

[project.scripts]
bheuu-scraper = "bheuu_scraper.cli:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-v --tb=short"

[tool.coverage.run]
source = ["src/bheuu_scraper"]
omit = ["tests/*"]
